backend-python/requirement.txt
Flask
SQLAlchemy
psycopg2-binary
redis
prometheus-client
python-dotenv
